#8: fix logic, restructure usecase
