+++ internal/usecase/group_usecase/group_usecase.go
@@ -1,28 +1,32 @@
-package usecase
+package group_usecase
 
 import (
+	"collector-telegram-bot/internal"
 	"collector-telegram-bot/internal/dto"
 	"collector-telegram-bot/internal/models"
 	repo "collector-telegram-bot/internal/repository"
+	"collector-telegram-bot/internal/usecase"
 	"fmt"
-	"github.com/sirupsen/logrus"
+	"github.com/google/uuid"
 )
 
-type GroupUsecase interface {
-	CreateSession(info dto.CreateSessionDTO) error
-	AddExpenseToSession(info dto.AddExpenseDTO) error
-}
+const (
+	ActiveSession = "active"
+	EmptyString   = ""
+)
 
 type AppGroupUsecase struct {
-	log  *logrus.Entry
+	log  internal.Logger
 	repo repo.Repository
 }
 
-func NewGroupUsecase(log *logrus.Entry, repo repo.Repository) GroupUsecase {
+func New(log internal.Logger, repo repo.Repository) GroupUsecase {
 	return &AppGroupUsecase{log: log, repo: repo}
 }
 
 func (uc *AppGroupUsecase) CreateSession(info dto.CreateSessionDTO) error {
+	sessionUUID := uuid.New()
+
 	userID, err := uc.upsertUser(info.UserID, info.Username)
 
 	if err != nil {
@@ -34,12 +38,12 @@ func (uc *AppGroupUsecase) CreateSession(info dto.CreateSessionDTO) error {
 	switch {
 	case err != nil:
 		return err
-	case curSession.UUID != 0:
-		return SessionExistsErr
+	case curSession.State == ActiveSession:
+		return usecase.SessionExistsErr
 	default:
 	}
 
-	session := models.NewSession(userID, info.ChatID, info.SessionName)
+	session := models.NewSession(sessionUUID, userID, info.ChatID, info.SessionName)
 	err = uc.repo.CreateNewSession(session)
 	if err != nil {
 		return fmt.Errorf("usecase: %v", err.Error())
@@ -51,7 +55,8 @@ func (uc *AppGroupUsecase) CreateSession(info dto.CreateSessionDTO) error {
 		return fmt.Errorf("usecase: %v", err.Error())
 	}
 	// Add creator to members
-	return uc.repo.AddMemberToSession(curSession.UUID, userID)
+	_, err = uc.repo.AddMemberToSession(curSession.UUID, userID)
+	return err
 }
 
 func (uc *AppGroupUsecase) upsertUser(userID int64, username string) (uint64, error) {
@@ -65,18 +70,11 @@ func (uc *AppGroupUsecase) upsertUser(userID int64, username string) (uint64, er
 	if user.ID == 0 {
 		user.Username = username
 		user.TgID = userID
-		err = uc.repo.CreateUser(user)
+		user.ID, err = uc.repo.CreateUser(user)
 		if err != nil {
 			return 0, fmt.Errorf("usecase: %v", err.Error())
 		}
 	}
-
-	// Check that user was created and get id from Database
-	user, err = uc.repo.GetUser(userID)
-	if err != nil {
-		return 0, fmt.Errorf("usecase: %v", err.Error())
-	}
-
 	return user.ID, nil
 }
 
@@ -88,8 +86,8 @@ func (uc *AppGroupUsecase) AddExpenseToSession(info dto.AddExpenseDTO) error {
 	}
 
 	// If session not exist -- return error
-	if session.UUID == 0 {
-		return SessionNotExistsErr
+	if session.State != ActiveSession {
+		return usecase.SessionNotExistsErr
 	}
 
 	// Check user is exists in db
@@ -108,13 +106,10 @@ func (uc *AppGroupUsecase) AddExpenseToSession(info dto.AddExpenseDTO) error {
 
 	// If user not in session, add as member
 	if memberID == 0 {
-		err = uc.repo.AddMemberToSession(session.UUID, userID)
+		memberID, err = uc.repo.AddMemberToSession(session.UUID, userID)
 		if err != nil {
 			return fmt.Errorf("usecase: %v", err.Error())
 		}
-		// Get UUID
-		member, _ = uc.repo.GetMemberBySession(session.UUID, userID)
-		memberID = member.ID
 	}
 
 	// Add user costs
