+++ internal/usecase/private_usecase/contract.go
@@ -1,1 +1,7 @@
 package private_usecase
+
+import tele "gopkg.in/telebot.v3"
+
+type PrivateUsecase interface {
+	Sessions(c tele.Context)
+}
