+++ internal/usecase/private_usecase/private_usecase.go
@@ -1,21 +1,17 @@
-package usecase
+package private_usecase
 
 import (
+	"collector-telegram-bot/internal"
 	"collector-telegram-bot/internal/repository"
-	"github.com/sirupsen/logrus"
 	tele "gopkg.in/telebot.v3"
 )
 
-type PrivateUsecase interface {
-	Sessions(c tele.Context)
-}
-
 type AppPrivateUsecase struct {
-	log  *logrus.Entry
+	log  internal.Logger
 	repo repo.Repository
 }
 
-func NewPrivateUsecase(log *logrus.Entry, repo repo.Repository) PrivateUsecase {
+func New(log internal.Logger, repo repo.Repository) PrivateUsecase {
 	return &AppPrivateUsecase{log: log, repo: repo}
 }
